[FLINK-14259] Introduce JobManagerRunner interface
This commit extracts the JobManagerRunner interface and renames the implementation
class into JobManagerRunnerImpl.

This closes #9788.
