+++ flink-runtime/src/main/java/org/apache/flink/runtime/dispatcher/DefaultJobManagerRunnerFactory.java
@@ -23,6 +23,7 @@ import org.apache.flink.runtime.heartbeat.HeartbeatServices;
 import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobmaster.JobManagerRunner;
+import org.apache.flink.runtime.jobmaster.JobManagerRunnerImpl;
 import org.apache.flink.runtime.jobmaster.JobManagerSharedServices;
 import org.apache.flink.runtime.jobmaster.JobMasterConfiguration;
 import org.apache.flink.runtime.jobmaster.factories.DefaultJobMasterServiceFactory;
@@ -39,7 +40,7 @@ import org.apache.flink.runtime.shuffle.ShuffleMaster;
 import org.apache.flink.runtime.shuffle.ShuffleServiceLoader;
 
 /**
- * Singleton default factory for {@link JobManagerRunner}.
+ * Singleton default factory for {@link JobManagerRunnerImpl}.
  */
 public enum DefaultJobManagerRunnerFactory implements JobManagerRunnerFactory {
 	INSTANCE;
@@ -75,7 +76,7 @@ public enum DefaultJobManagerRunnerFactory implements JobManagerRunnerFactory {
 			schedulerNGFactory,
 			shuffleMaster);
 
-		return new JobManagerRunner(
+		return new JobManagerRunnerImpl(
 			jobGraph,
 			jobMasterFactory,
 			highAvailabilityServices,
